+++ web/components/issues/issue-layouts/kanban/module-root.tsx
@@ -67,9 +67,14 @@ export const ModuleKanBanLayout: React.FC = observer(() => {
         moduleIssueStore.deleteIssue(group_by, sub_group_by, issue);
         issueDetailStore.deleteIssue(workspaceSlug.toString(), issue.project, issue.id);
       }
-      if (action === "remove") {
+      if (action === "remove" && issue.bridge_id) {
         moduleIssueStore.deleteIssue(group_by, null, issue);
-        moduleIssueStore.removeIssueFromModule(workspaceSlug.toString(), issue.project, moduleId.toString(), issue.id);
+        moduleIssueStore.removeIssueFromModule(
+          workspaceSlug.toString(),
+          issue.project,
+          moduleId.toString(),
+          issue.bridge_id
+        );
       }
     },
     [moduleIssueStore, issueDetailStore, moduleId, workspaceSlug]
